Refactored common types to shared file
