+++ pkg/apis/gcp/v1alpha1/common_types.go
@@ -1,0 +1,13 @@
+package v1alpha1
+
+// LegalEntity contains Red Hat specific identifiers to the original creator the clusters
+type LegalEntity struct {
+	Name string `json:"name"`
+	ID   string `json:"id"`
+}
+
+// NamespacedName contains the name of a secret and its namespace
+type NamespacedName struct {
+	Name      string `json:"name"`
+	Namespace string `json:"namespace"`
+}
